Revamp API Designer feature
