+++ composer/packages/api-designer/src/api-designer.ts
@@ -23,15 +23,12 @@ import OpenApiVisualizer from "./api-designer/index";
 
 export default function renderAPIEditor(renderTarget: HTMLElement,
                                         oasJson: string, onOASChange: () => void , updateView: boolean) {
-    let openApiJson = JSON.parse(oasJson);
-
-    if (updateView) {
-        openApiJson = JSON.parse(openApiJson);
-    }
+    const openApiJson = JSON.parse(oasJson);
 
     const oasVisualizer = createElement(OpenApiVisualizer, {
         onDidChange: onOASChange,
         openApiJson
     });
+
     render(oasVisualizer, renderTarget);
 }
