+++ composer/packages/api-designer/src/api-designer/components/info/license.tsx
