+++ composer/packages/api-designer/src/api-designer/operations/add-operation.tsx
@@ -83,7 +83,7 @@ class AddOpenApiOperation extends React.Component<OpenApiAddOperationProps, Open
                     this.setState({
                         operationMethods: []
                     });
-                }}>Save</Button>
+                }}>Add</Button>
             </Form>
         );
     }
