--- composer/packages/swagger-designer/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@ballerina/api-designer",
-  "version": "0.990.3-SNAPSHOT",
-  "description": "",
-  "main": "lib/index.js",
-  "scripts": {
-    "copy-less": "copyfiles -f src/**/*.less lib/api-designer/components/style/",
-    "prepublishOnly": "npm run build",
-    "prebuild": "npm run clean && npm run lint",
-    "clean": "composer pkg:clean",
-    "lint": "composer pkg:lint",
-    "build": "npm run copy-less && composer pkg:build",
-    "test": "echo \"No tests specified\"",
-    "coverage": "composer pkg:test:coverage",
-    "watch": "composer pkg:watch",
-    "storybook": "composer pkg:storybook"
-  },
-  "author": "ballerina.io",
-  "license": "Apache-2.0",
-  "devDependencies": {
-    "@ballerina/composer-cli": "^0.990.3-SNAPSHOT",
-    "copyfiles": "^2.1.0"
-  },
-  "dependencies": {
-    "@types/swagger-parser": "^4.0.2",
-    "js-yaml": "^3.12.0",
-    "openapi3-ts": "^1.1.0",
-    "react-markdown": "^4.0.4",
-    "semantic-ui-css": "^2.4.1",
-    "semantic-ui-react": "^0.83.0",
-    "swagger-parser": "^5.0.5",
-    "swagger-schema-official": "^2.0.0-bab6bed"
-  }
-}
