--- composer/packages/swagger-designer/src/api-designer.ts
@@ -1,34 +1,0 @@
-/**
- * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- *
- */
-
-import { createElement } from "react";
-import { render } from "react-dom";
-import OpenApiVisualizer from "./api-designer/index";
-
-export default function renderAPIEditor(renderTarget: HTMLElement,
-                                        oasJson: string, onOASChange: () => void , updateView: boolean) {
-    const openApiJson = JSON.parse(oasJson);
-
-    const oasVisualizer = createElement(OpenApiVisualizer, {
-        onDidChange: onOASChange,
-        openApiJson
-    });
-
-    render(oasVisualizer, renderTarget);
-}
