--- composer/packages/swagger-designer/tslint.json
@@ -1,10 +1,0 @@
-{
-    "extends": [
-        "../../configs/common.tslint.json"
-    ],
-    "linterOptions": {
-        "exclude": [
-            "**/*.json"
-        ]
-    }
-}
